opravy

+++ routes/web.php
@@ -2,6 +2,8 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\EventController;
+use App\Http\Controllers\Admin\EventController AS AdminEventController;
+use App\Http\Controllers\Admin\FileController AS AdminFileController;
 use App\Http\Controllers\FileController;
 
 /*
@@ -21,15 +23,24 @@ Route::get('/', function () {
 
 Route::controller(EventController::class)->group(function() {
     Route::get('udalost','index')->name('event.index');
-    Route::get('udalost/nova', 'new')->name('event.new');
     Route::get('udalost/{event}','show')->name('event.detail');
-    Route::get('udalost/{event}/editovat','edit')->name('event.edit');
-    Route::post('udalost', 'store')->name('event.store');
-    Route::put('udalost/{event}','update')->name('event.update');
-    Route::delete('udalost/{event}','delete')->name('event.delete');
+});
+
+Route::controller(AdminEventController::class)->group(function() {
+    Route::get('admin/udalost','index')->name('admin.event.index');
+    Route::get('admin/udalost/nova', 'new')->name('admin.event.new');
+    Route::get('admin/udalost/{event}','show')->name('admin.event.detail');
+    Route::get('admin/udalost/{event}/editovat','edit')->name('admin.event.edit');
+    Route::post('admin/udalost', 'store')->name('admin.event.store');
+    Route::put('admin/udalost/{event}','update')->name('admin.event.update');
+    Route::delete('admin/udalost/{event}','delete')->name('admin.event.delete');
 });
 
 Route::controller(FileController::class)->group(function() {
     Route::get('soubor/{file}/download','download')->name('file.download');
     Route::delete('soubor/{file}/smazat','delete')->name('file.delete');
+});
+
+Route::controller(AdminFileController::class)->group(function() {
+    Route::delete('soubor/{file}/smazat','delete')->name('admin.file.delete');
 });
